Explain post-commit hook for speaking commit title
In addition to the zsh script described previously in the readme, one can also add a post-commit hook in git to read the commit title aloud. This will work in bash, zsh, or whatever other shell. Its main limitation is that it's repo-specific, but I linked to info on how to work around/with that.